Ignore hidden files
This is a handy function to get the list of all variables in a folder, without hidden files (starting with a dot, e.g. .DS_store).

Use:

clean_list_of_pwd = nohiddenfiles()

to get the list of the pwd (as default)

Alternatively you can enter a path to another directory.